feat: create newsletter interface
